scripts/requirements.txt
python-docx>=1.1.0
docxtpl>=0.16.0
lxml>=5.0.0
pdfplumber>=0.11
camelot-py[cv]>=0.11
pymupdf>=1.25
Pillow>=10.0
matplotlib>=3.8
numpy>=1.26
jinja2>=3.1
